# static/swagger.json
{
  "swagger": "2.0",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "info": {
    "title": "Book API",
    "version": "1.0.0"
  },
  "paths": {
    "/register": {
      "post":{
        "security": [],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {},
          "400": {},
          "500": {}
        }
      }
    },
    "/login":{
      "post":{
        "security": [],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {},
          "400": {},
          "401": {},
          "500":{}
        }
      }
    },
    "/books": {
      "get": {
        "parameters": [
          {
            "name": "currentPage",
            "in": "query",
            "type": "integer",
            "required": true
          },
          {
            "name": "rowsPerPage",
            "in": "query",
            "type": "integer",
            "required": true
          },
          {
            "name": "orderBy",
            "in": "query",
            "type": "intstringeger",
            "required": false,
            "description": "Order By asc or desc "
          },
          {
            "name": "orderAttribute",
            "in": "query",
            "type": "string",
            "required": false,
            "description": "Order Attribute id or price "
          },
          {
            "name": "startDate",
            "in": "query",
            "type": "string",
            "required": false,
            "format": "date",
            "description": "Date in YYYY-MM-DD format"
          },
          {
            "name": "endDate",
            "in": "query",
            "type": "string",
            "required": false,
            "format": "date",
            "description": "Date in YYYY-MM-DD format"
          }
        ],
        "responses": {
          "200": {},
          "404": {},
          "500": {}
        }
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "isbn": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "author_id": {
                  "type": "integer"
                }
              },
              "required": [
                "title",
                "isbn",
                "price",
                "author_id"
              ]
            }
          }
        ],
        "responses": {
          "201": {},
          "400": {},
          "500": {}
        }
      }
    },
    "/books/search": {
      "get": {
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {},
          "404": {},
          "500": {}
        }
      }
    },
    "/books/author/{author_id}": {
      "get": {
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "currentPage",
            "in": "query",
            "type": "integer",
            "required": true
          },
          {
            "name": "rowsPerPage",
            "in": "query",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {},
          "404": {},
          "500": {}
        }
      }
    },
    "/books/{book_id}": {
      "get": {
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {},
          "404": {},
          "500": {}
        }
      },
      "put": {
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "isbn": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "author_id": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {},
          "400": {},
          "404": {},
          "500": {}
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {},
          "404": {},
          "500": {}
        }
      }
    },
    "/authors": {
      "get":{
        "responses": {
          "201": {},
          "400": {},
          "500": {}
        }
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "201": {},
          "400": {},
          "500": {}
        }
      }
    },
    "/authors/{author_id}":{
      "delete": {
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {},
          "404": {},
          "500": {}
        }
      }
    }
  }
}